vk: Refactor descriptors a bit
